refactor(claude-agent): split provider ownership (plan 068)
