+++ apps/server/src/modules/provider-contracts/provider-base.ts
@@ -71,7 +71,6 @@ export const ClaudeAgentConfigSchema = BaseProviderConfig.extend({
     }).optional(),
     subagentModel: ClaudeAgentModelEnvValueSchema.optional(),
   }).optional(),
-  permissionMode: z.enum(['bypassPermissions', 'plan']).default('bypassPermissions'),
   effort: z.enum(['low', 'medium', 'high', 'xhigh', 'max']).default('high'),
   allowDangerouslySkipPermissions: z.boolean().optional(),
   skills: z.union([z.literal('all'), z.array(z.string())]).optional(),
@@ -227,7 +226,6 @@ export function readTrustedClaudeAgentConfig(raw: string): ClaudeAgentConfig {
     skillPaths: config.skillPaths ?? [],
     additionalDirectories: config.additionalDirectories ?? [],
     claudeAgent: config.claudeAgent,
-    permissionMode: config.permissionMode === 'plan' ? 'plan' : 'bypassPermissions',
     effort: config.effort ?? 'high',
     allowDangerouslySkipPermissions: config.allowDangerouslySkipPermissions,
     skills: config.skills ?? [],
